Added XDG Base Directory support (#103)
* feat(config.go): xdg base directory support

Added support for XDG Base Directory.  If the XDG_CONFIG_HOME environment variable is set, look in the XDG_CONFIG_HOME/commitizen directory for the .czrc configuration file.  The order of operation is, look in the current repo directory, then user's home directory, and the XDG_CONFIG_HOME/commitizen directory.  If no configuration file is found, look in ~/.config/commitizen.

* Updated README.md to reflect previous changes.

---------

Co-authored-by: Jeff Lanzarotta <jlanzarotta@matw.com>